data grand lyon: refactor with base entity
