+++ homeassistant/components/data_grand_lyon/sensor.py
@@ -14,13 +14,12 @@ from homeassistant.components.sensor import (
 )
 from homeassistant.config_entries import ConfigSubentry
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 from homeassistant.helpers.typing import StateType
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import DOMAIN, SUBENTRY_TYPE_STOP, SUBENTRY_TYPE_VELOV_STATION
+from .const import SUBENTRY_TYPE_STOP, SUBENTRY_TYPE_VELOV_STATION
 from .coordinator import DataGrandLyonConfigEntry, DataGrandLyonCoordinator
+from .entity import DataGrandLyonEntity, DataGrandLyonVelovEntity
 
 PARALLEL_UPDATES = 0
 
@@ -192,12 +191,9 @@ async def async_setup_entry(
         )
 
 
-class DataGrandLyonStopSensor(
-    CoordinatorEntity[DataGrandLyonCoordinator], SensorEntity
-):
+class DataGrandLyonStopSensor(DataGrandLyonEntity, SensorEntity):
     """Sensor for Data Grand Lyon stop departures."""
 
-    _attr_has_entity_name = True
     entity_description: DataGrandLyonStopSensorEntityDescription
 
     def __init__(
@@ -207,19 +203,7 @@ class DataGrandLyonStopSensor(
         description: DataGrandLyonStopSensorEntityDescription,
     ) -> None:
         """Initialize the sensor."""
-        super().__init__(coordinator)
-        self.entity_description = description
-        self._subentry_id = subentry.subentry_id
-        assert subentry.unique_id is not None
-
-        self._attr_unique_id = f"{subentry.unique_id}-{description.key}"
-        self._attr_device_info = DeviceInfo(
-            identifiers={(DOMAIN, subentry.unique_id)},
-            name=subentry.title,
-            manufacturer="TCL",
-            model="Stop",
-            entry_type=DeviceEntryType.SERVICE,
-        )
+        super().__init__(coordinator, subentry, description, "TCL", "Stop")
 
     def _get_departure(self) -> TclPassage | None:
         """Return the departure for this sensor's index, or None."""
@@ -238,12 +222,9 @@ class DataGrandLyonStopSensor(
         return self.entity_description.value_fn(departure)
 
 
-class DataGrandLyonVelovSensor(
-    CoordinatorEntity[DataGrandLyonCoordinator], SensorEntity
-):
+class DataGrandLyonVelovSensor(DataGrandLyonVelovEntity, SensorEntity):
     """Sensor for Data Grand Lyon Vélo'v station."""
 
-    _attr_has_entity_name = True
     entity_description: DataGrandLyonVelovSensorEntityDescription
 
     def __init__(
@@ -253,27 +234,7 @@ class DataGrandLyonVelovSensor(
         description: DataGrandLyonVelovSensorEntityDescription,
     ) -> None:
         """Initialize the sensor."""
-        super().__init__(coordinator)
-        self.entity_description = description
-        self._subentry_id = subentry.subentry_id
-        assert subentry.unique_id is not None
-
-        self._attr_unique_id = f"{subentry.unique_id}-{description.key}"
-        self._attr_device_info = DeviceInfo(
-            identifiers={(DOMAIN, subentry.unique_id)},
-            name=subentry.title,
-            manufacturer="JCDecaux",
-            model="Station",
-            entry_type=DeviceEntryType.SERVICE,
-        )
-
-    @property
-    def available(self) -> bool:
-        """Return True if the station data is available."""
-        return (
-            super().available
-            and self._subentry_id in self.coordinator.data.velov_stations
-        )
+        super().__init__(coordinator, subentry, description)
 
     @property
     def native_value(self) -> StateType | datetime:
